Refactored to make the components more configurable and reusable
