+++ src/management-system-v2/app/(dashboard)/[environmentId]/settings/@processAutomation/page.tsx
@@ -1,90 +1,12 @@
-import { SettingGroup } from '../type-util';
-import SettingsSection from '../settings-section';
 import { env } from '@/lib/env-vars';
 import { populateSpaceSettingsGroup } from '@/lib/data/db/space-settings';
 import { getCurrentEnvironment } from '@/components/auth';
+import SettingsInjector from '../settings-injector';
+import { settings } from './settings';
+import Wrapper from './wrapper';
 
 const Page = async ({ params }: { params: { environmentId: string } }) => {
   if (!env.PROCEED_PUBLIC_ENABLE_EXECUTION) return null;
-
-  const settings: SettingGroup = {
-    key: 'process-automation',
-    name: 'Process Automation',
-    children: [
-      {
-        key: 'active',
-        name: 'Enabled',
-        type: 'boolean',
-        description: 'Controls whether this view is activated in this space.',
-        value: true,
-      },
-      {
-        key: 'tasklist',
-        name: 'Task List',
-        children: [
-          {
-            key: 'active',
-            name: 'Enabled',
-            type: 'boolean',
-            description: 'Controls whether this view is activated in this space.',
-            value: true,
-          },
-        ],
-      },
-      {
-        key: 'dashboard',
-        name: 'Dashboard',
-        children: [
-          {
-            key: 'active',
-            name: 'Enabled',
-            type: 'boolean',
-            description: 'Controls whether this view is activated in this space.',
-            value: true,
-          },
-        ],
-      },
-      {
-        key: 'projects',
-        name: 'Projects',
-        children: [
-          {
-            key: 'active',
-            name: 'Enabled',
-            type: 'boolean',
-            description: 'Controls whether this view is activated in this space.',
-            value: true,
-          },
-        ],
-      },
-      {
-        key: 'executions',
-        name: 'Executions',
-        children: [
-          {
-            key: 'active',
-            name: 'Enabled',
-            type: 'boolean',
-            description: 'Controls whether this view is activated in this space.',
-            value: true,
-          },
-        ],
-      },
-      {
-        key: 'machines',
-        name: 'Machines',
-        children: [
-          {
-            key: 'active',
-            name: 'Enabled',
-            type: 'boolean',
-            description: 'Controls whether this view is activated in this space.',
-            value: true,
-          },
-        ],
-      },
-    ],
-  };
 
   const {
     ability,
@@ -93,7 +15,12 @@ const Page = async ({ params }: { params: { environmentId: string } }) => {
 
   await populateSpaceSettingsGroup(spaceId, settings, ability);
 
-  return <SettingsSection sectionName="processAutomation" group={settings} priority={900} />;
+  return (
+    <>
+      <SettingsInjector sectionName="processAutomation" group={settings} priority={900} />
+      <Wrapper group={settings} />
+    </>
+  );
 };
 
 export default Page;
