+++ src/management-system-v2/app/(dashboard)/[environmentId]/settings/@processAutomation/settings.ts
@@ -1,0 +1,78 @@
+export const settings = {
+  key: 'process-automation',
+  name: 'Process Automation',
+  children: [
+    {
+      key: 'active',
+      name: 'Enabled',
+      type: 'boolean',
+      description: 'Controls whether this view is activated in this space.',
+      value: true,
+    },
+    {
+      key: 'tasklist',
+      name: 'Task List',
+      children: [
+        {
+          key: 'active',
+          name: 'Enabled',
+          type: 'boolean',
+          description: 'Controls whether this view is activated in this space.',
+          value: true,
+        },
+      ],
+    },
+    {
+      key: 'dashboard',
+      name: 'Dashboard',
+      children: [
+        {
+          key: 'active',
+          name: 'Enabled',
+          type: 'boolean',
+          description: 'Controls whether this view is activated in this space.',
+          value: true,
+        },
+      ],
+    },
+    {
+      key: 'projects',
+      name: 'Projects',
+      children: [
+        {
+          key: 'active',
+          name: 'Enabled',
+          type: 'boolean',
+          description: 'Controls whether this view is activated in this space.',
+          value: true,
+        },
+      ],
+    },
+    {
+      key: 'executions',
+      name: 'Executions',
+      children: [
+        {
+          key: 'active',
+          name: 'Enabled',
+          type: 'boolean',
+          description: 'Controls whether this view is activated in this space.',
+          value: true,
+        },
+      ],
+    },
+    {
+      key: 'machines',
+      name: 'Machines',
+      children: [
+        {
+          key: 'active',
+          name: 'Enabled',
+          type: 'boolean',
+          description: 'Controls whether this view is activated in this space.',
+          value: true,
+        },
+      ],
+    },
+  ],
+};
