+++ src/management-system-v2/app/(dashboard)/[environmentId]/settings/components.module.scss
